bench: move out MemHog into its own submodule, no functional change
